fix(lens): co-locate the processMemory cell store + equals with its definition
Move memoryCellStore, processMemoryMbEqual and its rssMb helper out of memorySampler.ts and into surface.ts beside the processMemory cell entry (mirroring terminalList), and re-home the equals unit test to surface.test.ts; the sampler now only owns the read+publish cadence.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-3, raised by hickey). Not pushed or merged.
